implement test_to_period (#32270)
